Fixed the sqlalchemy.exc.OperationalError: (MySQLdb.OperationalError) Unknown host bug, unknown password error, renamed files, etc
