--- backend/app/models/base_model.py
@@ -1,6 +1,0 @@
-from app.extensions import db
-from app.utils import generate_password_hash
-
-class Base(db.Model):
-    __abstract__ = True
-    id = db.Column(db.Integer, primary_key=True)
