--- backend/app/utils.py
@@ -1,4 +1,0 @@
-import bcrypt
-
-def generate_password_hash(password):
-    return bcrypt.hashpw(password.encode(), bcrypt.gensalt()).decode()
