+++ backend/diy_app/models/base_model.py
@@ -1,0 +1,5 @@
+from diy_app.extensions import db
+
+class Base(db.Model):
+    __abstract__ = True
+    id = db.Column(db.Integer, primary_key=True, autoincrement=True)
