+++ backend/mysql_setup.sql
@@ -3,7 +3,7 @@
 CREATE DATABASE IF NOT EXISTS diy_creative_display;
 
 -- If the user does not exist, create it
-CREATE USER IF NOT EXISTS 'diy_cd_user'@'localhost' IDENTIFIED BY 'DIY_@cd_pwd3}';
+CREATE USER IF NOT EXISTS 'diy_cd_user'@'localhost' IDENTIFIED BY 'DIY_cd_pwd3}';
 
 -- Grant privileges to the user for the database
 GRANT ALL PRIVILEGES ON diy_creative_display.* TO 'diy_cd_user'@'localhost';
@@ -19,7 +19,7 @@ CREATE TABLE IF NOT EXISTS users (
     id INT AUTO_INCREMENT PRIMARY KEY,
     name VARCHAR(50) NOT NULL,
     username VARCHAR(30) NOT NULL,
-    email VARCHAR(100) NOT NULL UNIQUE
+    email VARCHAR(100) NOT NULL UNIQUE,
 );
 
 -- Flush privileges to apply changes
